Test Exasol 6, too
Thanks to @ittleK0i for pointing out that Exasol 6 uses port 8888. I tried to
keep the single port spec in the service section, but it GitHub actions didn't
like it. I filed actions/runner#1508 to see what can be done. Also tried only
using the 7.x client libraries, which worked for 6.2 but never finished on 6.1
or 6.0. So keeping the use of the 6.x clients to test against 6.x.