Use unique registries during testing
To avoid conflicts when multiple tests run in parallel, as they sometimes do
in cpantesters smoketest configurations. First noticed for Snowflake and fixdd
in 8ba7070, so the this commit generalizes on that solution.

Only Oracle still tests static schema names, because the usernames are defined
in advance, not created at runtime.