Fix a Firebird test and upgrade issues
Commit bdc2b0caf conflated the target database with the registry database. So
give the registry a more distinctive name, restore the separate name for the
database we create for testing (when none exists already), and skip cleanup
for any database files that don't exist (because no live tests were run).

Also, disconnect from a local Firebird connection on upgrade. With a direct
connection (with no host name), there can be only one connection do the database
at a time. This is a problem for upgrades, where we already have a DBI
connection and then use `isql` to perform the upgrade. So disconnect DBI from
the database before performing an upgrade, and add a Moo method to clear the
`dbh` attribute, which will allow it to reconnect the next time it's called.

Figured out by using the Docker image specified in #597 and installing the
`firebird3.0-utils` package.

Fixes #597.