support for YugabyteDB
Works great with the Postgres engine except for locks. So detect that it's
Yugabyte and don't create the locks if it is. Change the handling of the
`lock_sql` option in the database tests to take a code reference instead of a
hash reference, and only return the locking instructions in `t/pg.t` if
the engine is not Yugabyte. Mention the Yugabyte support here and there.

Alway sort the current state test tags for Yugabyte, as previously done only
for Oracle and MySQL. The Postgres aggregate docs say that the ordering of
`array_agg()` is undefined, though it has been reliable in all testing. Not so
for Yugabyte. So we force the sort for Yugabyte, but also added #636 to
propose a better path forward for aggregate sorting

Add GitHub Actions testing for Yugabyte 2.6 and higher, and link it in the
README. While at it, clean up the README, properly listing all supported
engines and linking to them.