Lock destination after registry creation
Turns out that v1.2.0 introduced a bug when trying to deploy to a MySQL server
that does not already have a registry database (usually named "sqitch"). It
would fail because it could not connect to the registry. Which makes sense,
since on first deploy it does not usually exist yet.

So initialize the MySQL registry if it does not exist, and move revert's call
to lock the destination to after we're certain the registry exists. Since this
might increase the calls to `initialized()`, change it to an attribute to cut
down on the database calls.

Resolves #601.