Improve the default client search for sqlplus
Some installations of SQL*Plus live in `$ORACLE_HOME/bin` rather than
just in `$ORACLE_HOME`. So refine `default_client()` to search for an
executable file in `$ORACLE_HOME/bin` and return that value if it
exists, then look for an executable file in `$ORACLE_HOME` and return it
if it exists. Otherwise simply return `sqlplus`. This changes the
previous behavior where it would always return `$ORACLE_HOME/sqlplus` if
`$ORACLE_HOME` is set. Now it will fall back on a path search if
`sqlplus` isn't in `ORACLE_HOME` even when `ORACLE_HOME` is set.
Resolves #747.