Remove identifier quoting from snowflake
Added in v1.4.0. Turns out Snowflake allows a warehouse to be specified
in a different database, in which case dots are valid in the name and
should not be quoted! So users must properly quote when necessary, but
added notes to `sqitchtutorial-snowflake.pod` on the need to use URI
escapes for special characters. Resolves #685 (again/for real).

Also add a brief discussion on URL encoding to `sqitch-target.pod`, the
main reference for database URLs in the Sqitch documentation.