Fix multi-engine template resolution
When adding a change to multiple engines, Sqitch would randomly pick one
engine's templates or the others and create them all for the same
engine. In other words, adding a change to both "sqlite" and "pg"
targets would result in all change files generated from templates from
only one of those engines -- whichever it happened to find first.

Due to the use of a hash reference for templates in add.pm. Fix it by
making a copy of the hash instead. This leads to duplication of effort
when multiple targets use the same engine, but the cost is generally low
relative to how often `add` is called -- and likely is measured in
milliseconds.

Resolves #795.