Use locale pragma instead of POSIX::setlocale()
The Pragma is more likely to do the right thing, as confirmed by new
tests, which fail when using `setlocale` and now succeed with `use
locale`. The tests, in `xt/locale`, include compiled locale dictionaries
(`*.mo` files) with a single message for the tested languages. This is
in contrast to the released locale dictionaries, which are generated at
release time but not stored in the repository.

Update the `Language-Team` header in the project localization packages
in `po` directory to `Sqitch Hackers <sqitch-hackers@googlegroups.com>`.

Update the `os.yml` and `perl.yml` workflows, which run all tests
including the new locale tests, to install the required locales on Linux
and to set the full `runs-on:` image name in the matrix (in response to
shogo82148/actions-setup-perl#1699).

Also remove the installation of an older version of Locale::TextDomain
from those workflows, since gflohr/libintl-perl#7 has been fixed and
released.

While at it, upgrade to `actions/checkout@v4` in all workflows and use
`runner.os` instead of `matrix.os` in conditionals.