Properly avoid advisory Yugabyte lock errors
The `try_lock` method was not properly checking for Yugabyte before
attempting to acquire an advisory lock. Likely it used to be a warning,
but now is an error that cannot be avoided. So have `try_lock` check for
Yugabyte before attempting to take an advisory lock. This requires
fetching the database handle, first, as the database is not identified
until it connects. So abstract that need by replacing the `_provider`
attribute with a method that checks a DBI private module attribute.
Fixes #841.

While at it, replace a wayward mention of the long-unsupported (or maybe
incorrectly documented?) `SQITCH_URI` environment variable in the Oracle
tutorial with the correct name, `SQITCH_TARGET`. Resolves #845.