Improve unknown engine error and error details
Add the `details_string` method to App::Sqitch::X. It includes just the
previous error and the stack trace. Then teach the `go` method in
`App::Sqitch` to use it for unexpected errors, and to use `as_string`
for unexpected DEV errors. This adds additional context to error output
for DEV errors and for other errors when trace mode is enabled (by
`-vvv`).

Then change the error loading an engine package from a DEV error to an
engine error, and provide a localized error message. This results in the
much less noisy error "Unknown engine: foo" when a user specifies an
unknown engine URI, such as `db:foo:`. Fixes #838.