Properly handle MySQL URIs without a DB name
The MySQL engine has in theory always allowed a URI without a database
name, though it emits a warning. But the `mysql` CLI option handling did
not properly account for it, expecting to always find a `--database`
option to replace.

Fix it by replacing the `--database` option when it exists and appending
it when it doesn't. Fixes #821.