Omit MySQL checkit() when it cannot be created
Remove the use of the `checkit()` function from the MySQL triggers that
enforce constraints on the `dependencies` table, then modify the
deployment of MySQL registry scripts to remove `checkit()` when MySQL is
using binary logging and Sqitch is not a super user. Fixes #824, which
reports that MySQL with binary logging enabled disallows the creation of
`DETERMINISTIC` functions by non-super users. In such a configuration,
Sqitch itself functions exactly as before, but `checkit()` is not
available for use in verify scripts.

Refactor the MySQL engine's `run_upgrade()` method to move the
rejiggering of a file to a new method, `_prepare_registry_file`. Teach
that method then to remove bits of the script as appropriate to the
MySQL version and permissions.

Update the MySQL tutorial to note how to add `checkit()` if it's
missing.