Skip registry creation on Snowflake when necessary
In some cases, the Sqitch user will lack permission to create a schema,
such as when the DBA doesn't want to give it that permission. Implement
`run_upgrade` in the Snowflake engine to detect this lack of permission
by executing `CREATE SCHEMA IF NOT EXISTS` and stripping out `CREATE
SCHEMA` and `COMMENT ON SCHEMA` commands when it fails due to
insufficient privileges. Resolves #828.