Extract checkit function from mysql.sql
The permission check added in 93a0d1f (#824) does not work if the
current user does not have permission to access the `mysql.user` table.
So remove the permission check and move the creation of the `checkit`
function to mysql.pm, where it ignores the failure to create the
function due to a lack of sufficient permission or binary logging.
Resolves #874.