Use Snowflake key pair auth, quote identifiers
Switch to key pair authentication in the Snowflake workflows, as
password-only authentication will be discontinued later this year. In
the process, work out and document how to set the appropriate ODBC
parameters in an `odbc.ini` file in preference to the connection URL.

Refactor `snowflake.sh` to∑ set `WORKSPACE` when installing SnowSQL,
which is required to upgrade to v1.3. Use the variable instead of
heard-coded paths throughout, and fix the location of the configuration
file (it has to go into the `.snowsql` subdirectory of the workspace).
Copy and update an `odbc.ini` and `.snowsql/config` as needed for key
pair auth.

Fix the quoting of the role and schema names on connecting to Snowflake,
which were silently failing and therefore failing to properly use the
registry schema, which lead to a failure to find the registry. Broken in
33291bd (#853).

While at it, remove support for the `SNOWSQL_PORT` environment variable,
which has long been deprecated by Snowflake and likely never did
anything. Also add redaction of passwords from Snowflake URL query
parameters in the display URL. Any query parameter matching `pwd` will
now be shown as "REDACTED".