Fix unwrapped Firebird connection errors
As revealed via cpan-testers, detect when `t/firebird.t` fails to raise
an App::Sqitch::X error and promote it to one. This allows DBIEngineTest
to properly display the error without dying. While at it, tweak
DBIEngineTest to always show diagnostics for connection failures,
whether or not a live test is required.