[#839] Adjusts permissions of `sqitch.conf` (and friends) every time
Every time the configuration is loaded, the file permissions are
checked and if they are different from `0600`, they are adjusted via
an explicit `chmod`.
The rationale is to ensure, as much as possibile, that configuration
including target credentials remains private to the user.

This introduces the dependency with `File::stat`, in order to perform
a more OOP-`stat` call.
The `App::Sqitch::Config::load_file` method has changed in order to
call a new function, `::chmod_files_if_needed` that accepts a list of
file-ish stuff to check against permissions.