Factor out array splitting
Always require arrays to be returned as arrays. Move inline splitting on
blank spaces with a new method, `_parse_array`, so that it applies more
consistently. Use it also in DBIEngineTest when fetching records, and
format lists of tags as arrays to compare rather than whatever the
database happens to return. This ensures that engines that don't support
arrays, like SQLite, behave consistently with array-supporting engines.