Get all tests passing for ClickHouse
Not too many workarounds the most significant workaround is to parse
array values, since DBD::ODBC doesn't yet support them. DBIEngine was
already doing some parsing for engines without arrays, splitting on
whitespace. So add the `_parse_array` method and override it in the
ClickHouse engine. Fortunately, ClickHouse arrays appear to be
syntactically equivalent to Perl arrays, so it just `eval`s the arrays.
A little hinky, but simple. This workaround can be ripped out once
clickHouse/clickhouse-odbc#525 is fixed or some other driver ships that
better integrates with the DBI.

Other changes:

*   Turn on `odbc_utf8_on` so we get proper Unicode encoding decoding
*   Always use literals in OFFSET and LIMIT
*   Override `log_revert_change`to select arrays rather than delimited
    strings
*   Work around ClickHouse returning default values rather than `NULL`s
    from empty CTEs in `changes_requiring_change` and
    `name_for_change_id`. The former also requires explicit column
    names, otherwise it includes the table aliases in the names.
*   Override `log_new_tags` to query for existing tags before inserting
    new ones. The `INSERT` with a `JOIN` did not work in ClickHouse.
*   ClickHouse requires a `WHERE` clause for `DELETE` and `UPDATE`
    expressions, so throw in `WHERE `1=1` where the test suite deletes
    or updates all rows.