Add NativePort, ClickHouse query params, tests
Add processing of variables to the ClickHouse `cli` method. This allows
one to use variables in ClickHouse client scripts.

Add a custom URL query parameter, `NativePort`, because the CLI uses the
ClickHouse "native" port, while the ODBC driver uses an HTTP port. If
both are using the defaults (9000 or 9440 for the CLI and 8123 or 8443
for ODBC) it all works, but if the HTTP port is different, only ODBC can
use it. So in addition to the HTTP port specified in the `db:clickhouse`
URL, which impacts only ODBC, the `NativePort` param only impacts the
CLI.

Other changes:

*   Fix a failing test by adding `clickhouse` to the list of engines in
    App::Sqitch::Command
*   Fix a failing test on windows by accounting CLI executable
    extensions (e.g., `.exe`)
*   Add tests for ClickHouse `_parse_array` and fix an issue with
    encoding
*   Add a few more tests to reach 100% code coverage
*   Remove unneeded fractional seconds copied from `mysql.t` and not
    needed for ClickHouse
*   Add a couple new terms to the spelling list in
    `xt/release/pod-spelling.t`