Add support for ClickHouse client config files
Follow as closely as possible the process for merging configurations
used by the ClickHouse CLI itself. That means:

*   Searching for the appropriate configuration files and stopping on
    the first one found
*   Supporting both XML and YAML configuration file formats
*   Merging host-specific configuration into root-level configuration

Use YAML::Tiny and XML::Tiny to handle the parsing of both formats. Use
a corpus of example configurations from the ClickHouse source code and
documentation to ensure both work. It turns out that installing
XML::LibXML is not easy on Windows, but XML::Tiny works without issue,
so go with it.

Use the configuration to update the deployment URL, username, and
password as appropriate. Enforce a secure connection if the
configuration file calls for it, uses the encrypted port, or specifies a
ClickHouse cloud domain. Map configuration variables to ODBC parameters
where equivalents exist.

The reverse isn't supported, alas, as the ClickHouse CLI doesn't have
options or environment variables for TLS client config, and it doesn't
seem sensible to try to write one. Better to work toward getting better
option support into the CLI.

This support for reading ClickHouse client configuration allows Sqitch
to support TLS client authentication. Update the `sqitch-authentication`
to document this support.

Move most instances of home directory detection to App::Sqitch::Config
to reduce duplication and to allow it to be mocked for testing.

While at it, remove defaulting the username to the system username, as
ClickHouse does not work that way, instead falling back on "default"
when no username is provided.