Search for `clickhouse` and `clickhouse-client`
Bare `clickhouse` is the canonical name, and `clickhouse-client` should
generally be a symlink, but some packages, like the [clickhouse-client
Apt package], just have `clickhouse-client`. This is a bad assumption
and should not be the default.

So search the path for both and return the first one found. Raise an
error if neither is found. Adopt and generalize the error message
previously used by the Firebird engine.

In passing, update the error regular expression to match more generally.
Previously it matched unixODBC-specific errors, which did not work on
macOS using the [clickhouse-odbc Homebrew formula], which uses iODBC. So
switch to an error message that comes only from the ClickHouse server.

  [clickhouse-client Apt package]: https://packages.debian.org/bullseye/clickhouse-client
  [clickhouse-odbc Homebrew formula]: https://formulae.brew.sh/formula/clickhouse-odbc#default